+++ AOS/app/src/main/java/com/example/eeos/presentation/topappbar/MemberStatusButtons.kt
@@ -1,4 +1,4 @@
-package com.example.eeos.presentation.common
+package com.example.eeos.presentation.topappbar
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.layout.Column
